Started work on other buttons and cleaned up code
I've started work on making other buttons and also I've cleaned up the
code so that ti would be easier to use and so that I don't have to keep
giving each div the same properties. Also, I removed "the useless
button" from "a special place" and added it to a new place: "the button
section".