Added main page, a new section, google verification and jQuery source code.
I’ve decided to rework the workings of the site and made a home page
for when the user starts up the site. Then, there will be multiple
pages branching out, like beginnings, which I’ve finished work on and
is the original, and evolution, the page I’m working on now which will
contain JavaScript code. Also, I added the jQuery source code so that I
can start using jQuery. I’ve also added a google verification
certificate for the website which I can use for when people google the
site.